style: refactor models
