+++ backend/api/models/__init__.py
@@ -1,0 +1,3 @@
+from .courseModel import *  # noqa: W0611
+from .exerciseModel import *  # noqa: W0611
+from .sessionModel import *  # noqa: W0611
